Initial commit
For easier access to the flowcharts and diagrams, I will introduce
Graphviz into the project tentatively. The image file will be able to
provide a very quick image of our diagrams and will remove any hassle of
opening other links. The gv file holds the code which produced the png
file.